+++ express-zod-api/src/index.ts
@@ -21,6 +21,7 @@ export { ServeStatic } from "./serve-static";
 export { createServer, attachRouting } from "./server";
 export {
   DocumentationError, // @todo remove in v30
+  EndpointResponseError,
   RoutingError,
   OutputValidationError,
   InputValidationError,